refactor(cmd): delete db.go, move helpers to serve.go
loadDBPath and resolveConfigPath are only used by serve.go in
config-file mode. Move them there and delete the now-empty db.go.
